Added module for book instance class
